ft(project): project issue - move project issue inside project - refactor projectIssue Controller [Finishes]
