+++ resources/views/projects/partials/risk-management/partials/add.blade.php
