+++ resources/views/projects/partials/risk-management/partials/resolved.blade.php
