+++ routes/web.php
@@ -99,6 +99,7 @@ Route::prefix('projects')->middleware(['auth'])->group(function () {
     Route::get('quotation/{project}', [ProjectController::class, 'quotation'])->name('project.quotation');
     Route::get('customer_po/{project}', [ProjectController::class, 'customerPo'])->name('project.customer-po');
     Route::get('invoice/{project}', [ProjectController::class, 'invoice'])->name('project.invoice');
+    Route::get('issues/{project}', [ProjectController::class, 'issues'])->name('project.issues');
     Route::post('', [ProjectController::class, 'store'])->name('project.store');
     Route::put('{project}', [ProjectController::class, 'update'])->name('project.update');
     Route::delete('{project}', [ProjectController::class, 'destroy'])->name('project.delete');
